# Optimize article_ids query by batching titles to prevent large IN clause slowdowns