--- a/app/models/wiki_content/category.rb
+++ b/app/models/wiki_content/category.rb
@@ -81,7 +81,7 @@ class Category < ApplicationRecord
     return @article_ids if @article_ids.present?
 
     @article_ids = []
-    article_titles.each_slice(2000) do |titles_batch|
+    article_titles.each_slice(5000) do |titles_batch|
       @article_ids.concat(Article.where(namespace: 0, wiki_id:, title: titles_batch).pluck(:id))
     end