@gabina 2025-07-15
I like this idea. Are you estimating previous query execution time vs current query execution time to see the improvement? What about memory?