@Abishekcs 2025-07-16
> Are you estimating previous query execution time vs current query execution time to see the improvement?

Yes these is one of the ways i am trying to measure the improvement.

<img width="512" height="387" alt="Screenshot from 2025-07-16 15-03-30" src="https://github.com/user-attachments/assets/c0e789c6-6da0-4659-ad5f-88b8fc5fa952" />


**this is the benchmark code:** 

```ruby

  def article_ids # rubocop:disable Metrics/MethodLength,Metrics/AbcSize
    require 'benchmark'

    # First approach - Single query
    articles_titles = Article.all.pluck(:title)

    single_query_time = Benchmark.measure do
      @article_ids = Article.where(namespace: 0, wiki_id:, title: articles_titles).pluck(:id)
    end

    # Second approach - Batched queries
    @article_ids = nil # Reset for fair comparison

    batched_query_time = Benchmark.measure do
      @article_ids = articles_titles.each_slice(500).flat_map do |titles_batch|
        Article.where(namespace: 0, wiki_id:, title: titles_batch).pluck(:id)
      end
    end

    puts 'Single Query Results:'
    puts single_query_time
    puts "Article IDs found: #{@article_ids.count}"
    puts '---'

    puts 'Batched Query Results:'
    puts batched_query_time
    puts "Article IDs found: #{@article_ids.count}"
    puts '---'

    # Compare the results
    puts 'Performance Comparison:'
    puts "Single query time: #{single_query_time.real.round(4)}s"
    puts "Batched query time: #{batched_query_time.real.round(4)}s"
    puts "Improvement: #{((single_query_time.real - batched_query_time.real) / single_query_time.real * 100).round(2)}%"
  end


```