@Abishekcs 2025-07-16
### **Update:**

I was able to reduce the memory consumption too with recent changes to the code:

``` ruby

  def article_ids
    return @article_ids if @article_ids.present?

    @article_ids = []
    article_titles.each_slice(500) do |titles_batch|
      @article_ids.concat(Article.where(namespace: 0, wiki_id:, title: titles_batch).pluck(:id))
    end

    @article_ids.uniq! || @article_ids
  end

```
**This code save memory by:** 

- concat appends directly to a single array (@article_ids).

- Only one main array is used throughout — no need for flattening or reallocation.

- concat is in-place and uses Array method under the hood, which is efficient.

- .uniq! mutates the array once `(i,e in the original variable it self)`, instead of copying.


### Benchmark Screenshot (Time + Memory)


<img width="357" height="263" alt="NEW 1" src="https://github.com/user-attachments/assets/e5ccc1e6-e5f1-4d8c-861e-20fee4178b53" />
<img width="357" height="263" alt="New 3" src="https://github.com/user-attachments/assets/289a01d7-53b9-4f21-851a-4d0f97f8da08" />
<img width="357" height="263" alt="New 2" src="https://github.com/user-attachments/assets/c6efac0b-9910-4deb-b907-30deb2c0ceef" />
<img width="357" height="263" alt="Screenshot from 2025-07-17 00-16-08" src="https://github.com/user-attachments/assets/1d3d0fc8-2ebf-4192-b426-935110612eb8" />


### Benchmark (Code)


```ruby

 def article_ids # rubocop:disable Metrics/MethodLength,Metrics/AbcSize
    require 'benchmark'

    # Simulated large input i.e take all the data available in my local db its around 155481
    articles_titles = Article.all.pluck(:title)

    puts "Benchmarking with #{articles_titles.size} titles..."

    results = {}

    ### Original (Single Query) ###
    GC.start(full_mark: true, immediate_sweep: true)
    ObjectSpace.trace_object_allocations_start

    memory_before = ObjectSpace.memsize_of_all
    time_single = Benchmark.realtime do
      @article_ids1 = Article.where(namespace: 0, wiki_id:, title: articles_titles).pluck(:id)
    end
    memory_after = ObjectSpace.memsize_of_all

    results[:single] = {
      time: time_single.round(3),
      memory_mb: ((memory_after - memory_before) / 1024.0 / 1024.0).round(2),
      count: @article_ids1.size
    }

    ### Batched Query ###
    GC.start(full_mark: true, immediate_sweep: true)
    ObjectSpace.trace_object_allocations_start

    memory_before = ObjectSpace.memsize_of_all
    time_batched = Benchmark.realtime do
      @article_ids2 = []
        articles_titles.each_slice(500) do |batch|
          @article_ids2.concat(Article.where(namespace: 0, wiki_id:, title: batch).pluck(:id))
        end
      @article_ids2.uniq!
    end
    memory_after = ObjectSpace.memsize_of_all

    results[:batched] = {
      time: time_batched.round(3),
      memory_mb: ((memory_after - memory_before) / 1024.0 / 1024.0).round(2),
      count: @article_ids2.size
    }

    ### Output ###
    puts "\n=== Results ==="
    results.each do |key, data|
      puts "#{key.capitalize} Query:"
      puts "  Time:        #{data[:time]}s"
      puts "  Memory used: #{data[:memory_mb]} MB"
      puts "  IDs fetched: #{data[:count]}"
    end

    if results[:batched][:time] < results[:single][:time]
      gain = ((results[:single][:time] - results[:batched][:time]) / results[:single][:time] * 100).round(2)
        puts "Batching is faster by #{gain}%"
    else
      loss = ((results[:batched][:time] - results[:single][:time]) / results[:batched][:time] * 100).round(2)
        puts "Batching is slower by #{loss}%"
    end
  end


```